+++ app/charts/column/columns-grouped-state.tsx
@@ -109,6 +109,7 @@ const useColumnsGroupedState = (
     allData,
   } = data;
   const { fields, interactiveFiltersConfig } = chartConfig;
+  const { y } = fields;
 
   const { width, height } = useSize();
   const formatNumber = useFormatNumber({ decimals: "auto" });
@@ -260,29 +261,35 @@ const useColumnsGroupedState = (
     const xScaleTimeRange = scaleTime().domain(xScaleTimeRangeDomain);
 
     // y
-    const minValue = getMinY(scalesData, (d) =>
-      getYErrorRange ? getYErrorRange(d)[0] : getY(d)
-    );
-    const maxValue = Math.max(
-      max(scalesData, (d) =>
-        getYErrorRange ? getYErrorRange(d)[1] : getY(d)
-      ) ?? 0,
-      0
-    );
-    const yScale = scaleLinear().domain([minValue, maxValue]).nice();
+    const yScale = scaleLinear();
+    const paddingYScale = scaleLinear();
 
-    const minPaddingValue = getMinY(paddingData, (d) =>
-      getYErrorRange ? getYErrorRange(d)[0] : getY(d)
-    );
-    const maxPaddingValue = Math.max(
-      max(paddingData, (d) =>
-        getYErrorRange ? getYErrorRange(d)[1] : getY(d)
-      ) ?? 0,
-      0
-    );
-    const paddingYScale = scaleLinear()
-      .domain([minPaddingValue, maxPaddingValue])
-      .nice();
+    if (y.customDomain) {
+      yScale.domain(y.customDomain);
+      paddingYScale.domain(y.customDomain);
+    } else {
+      const minValue = getMinY(scalesData, (d) => {
+        return getYErrorRange?.(d)[0] ?? getY(d);
+      });
+      const maxValue = Math.max(
+        max(scalesData, (d) => {
+          return getYErrorRange?.(d)[1] ?? getY(d);
+        }) ?? 0,
+        0
+      );
+      yScale.domain([minValue, maxValue]).nice();
+
+      const minPaddingValue = getMinY(paddingData, (d) => {
+        return getYErrorRange?.(d)[0] ?? getY(d);
+      });
+      const maxPaddingValue = Math.max(
+        max(paddingData, (d) => {
+          return getYErrorRange?.(d)[1] ?? getY(d);
+        }) ?? 0,
+        0
+      );
+      paddingYScale.domain([minPaddingValue, maxPaddingValue]).nice();
+    }
 
     return {
       colors,
@@ -316,6 +323,7 @@ const useColumnsGroupedState = (
     getYErrorRange,
     getY,
     getMinY,
+    y.customDomain,
   ]);
 
   // Group