+++ app/charts/line/lines-state.tsx
@@ -155,35 +155,49 @@ const useLinesState = (
   const xScaleTimeRange = scaleTime().domain(xScaleTimeRangeDomain);
 
   // y
-  const minValue = getMinY(scalesData, (d) =>
-    getYErrorRange ? getYErrorRange(d)[0] : getY(d)
-  );
-  const maxValue =
-    max(scalesData, (d) => (getYErrorRange ? getYErrorRange(d)[1] : getY(d))) ??
-    0;
-  const yDomain = [
-    minLimitValue !== undefined ? Math.min(minLimitValue, minValue) : minValue,
-    maxLimitValue !== undefined ? Math.max(maxLimitValue, maxValue) : maxValue,
-  ];
-  const yScale = scaleLinear().domain(yDomain).nice();
+  const yScale = scaleLinear();
+  const paddingYScale = scaleLinear();
 
-  const paddingMinValue = getMinY(paddingData, (d) =>
-    getYErrorRange ? getYErrorRange(d)[0] : getY(d)
-  );
-  const paddingMaxValue =
-    max(paddingData, (d) =>
-      getYErrorRange ? getYErrorRange(d)[1] : getY(d)
-    ) ?? 0;
-  const paddingYScale = scaleLinear()
-    .domain([
-      minLimitValue !== undefined
-        ? Math.min(minLimitValue, paddingMinValue)
-        : paddingMinValue,
-      maxLimitValue !== undefined
-        ? Math.max(maxLimitValue, paddingMaxValue)
-        : paddingMaxValue,
-    ])
-    .nice();
+  if (y.customDomain) {
+    yScale.domain(y.customDomain);
+    paddingYScale.domain(y.customDomain);
+  } else {
+    const minValue = getMinY(scalesData, (d) => {
+      return getYErrorRange?.(d)[0] ?? getY(d);
+    });
+    const maxValue =
+      max(scalesData, (d) => {
+        return getYErrorRange?.(d)[1] ?? getY(d);
+      }) ?? 0;
+    yScale
+      .domain([
+        minLimitValue !== undefined
+          ? Math.min(minLimitValue, minValue)
+          : minValue,
+        maxLimitValue !== undefined
+          ? Math.max(maxLimitValue, maxValue)
+          : maxValue,
+      ])
+      .nice();
+
+    const paddingMinValue = getMinY(paddingData, (d) => {
+      return getYErrorRange?.(d)[0] ?? getY(d);
+    });
+    const paddingMaxValue =
+      max(paddingData, (d) => {
+        return getYErrorRange?.(d)[1] ?? getY(d);
+      }) ?? 0;
+    paddingYScale
+      .domain([
+        minLimitValue !== undefined
+          ? Math.min(minLimitValue, paddingMinValue)
+          : paddingMinValue,
+        maxLimitValue !== undefined
+          ? Math.max(maxLimitValue, paddingMaxValue)
+          : paddingMaxValue,
+      ])
+      .nice();
+  }
 
   // segments
   const segmentFilter = segmentDimension?.id