+++ app/config-types.ts
@@ -330,6 +330,13 @@ export type UncertaintyFieldExtension = t.TypeOf<
   typeof UncertaintyFieldExtension
 >;
 
+const CustomScaleDomainFieldExtension = t.partial({
+  customDomain: t.tuple([t.number, t.number]),
+});
+export type CustomScaleDomainFieldExtension = t.TypeOf<
+  typeof CustomScaleDomainFieldExtension
+>;
+
 const ChartSubType = t.union([t.literal("stacked"), t.literal("grouped")]);
 export type ChartSubType = t.TypeOf<typeof ChartSubType>;
 
@@ -349,6 +356,7 @@ const ColumnFields = t.intersection([
       GenericField,
       ShowValuesFieldExtension,
       UncertaintyFieldExtension,
+      CustomScaleDomainFieldExtension,
     ]),
     color: t.union([SegmentColorField, SingleColorField]),
   }),
@@ -383,7 +391,11 @@ export type BarSegmentField = t.TypeOf<typeof BarSegmentField>;
 
 const BarFields = t.intersection([
   t.type({
-    x: t.intersection([GenericField, ShowValuesFieldExtension]),
+    x: t.intersection([
+      GenericField,
+      ShowValuesFieldExtension,
+      CustomScaleDomainFieldExtension,
+    ]),
     y: t.intersection([GenericField, SortingField]),
     color: t.union([SegmentColorField, SingleColorField]),
   }),
@@ -422,6 +434,7 @@ const LineFields = t.intersection([
       GenericField,
       ShowValuesFieldExtension,
       UncertaintyFieldExtension,
+      CustomScaleDomainFieldExtension,
       t.partial({
         showDots: t.boolean,
         showDotsSize: t.union([
@@ -474,6 +487,7 @@ const AreaFields = t.intersection([
     y: t.intersection([
       GenericField,
       ShowValuesFieldExtension,
+      CustomScaleDomainFieldExtension,
       t.partial({ imputationType: ImputationType }),
     ]),
     color: t.union([SegmentColorField, SingleColorField]),